move simple examples of rules to category packages
