+++ sail-point-customization/src/main/java/com/sailpoint/rule/connector/SimpleBuildMapRule.java
@@ -1,4 +1,4 @@
-package com.sailpoint.rule;
+package com.sailpoint.rule.connector;
 
 import com.sailpoint.annotation.Rule;
 import com.sailpoint.annotation.common.Argument;
